chore: Remove workspace root local dependencies
