+++ package.json
@@ -89,13 +89,14 @@
     "prepublishOnly": "turbo run build",
     "lint-staged": "lint-staged",
     "next-with-deps": "./scripts/next-with-deps.sh",
-    "next": "cross-env NEXT_PRIVATE_LOCAL_DEV=1 NEXT_TELEMETRY_DISABLED=1 NODE_OPTIONS=\"--trace-deprecation --enable-source-maps\" next",
+    "create-next-app": "packages/create-next-app/dist/index.js",
+    "next": "cross-env NEXT_PRIVATE_LOCAL_DEV=1 NEXT_TELEMETRY_DISABLED=1 NODE_OPTIONS=\"--trace-deprecation --enable-source-maps\" packages/next/dist/bin/next",
     "next-no-sourcemaps": "echo 'No longer supported. Use `pnpm next --disable-source-maps` instead'; exit 1;",
     "clean-trace-jaeger": "node scripts/rm.mjs test/integration/basic/.next && TRACE_TARGET=JAEGER pnpm next build test/integration/basic",
     "debug": "cross-env NEXT_PRIVATE_LOCAL_DEV=1 NEXT_TELEMETRY_DISABLED=1 node --inspect --trace-deprecation --enable-source-maps packages/next/dist/bin/next",
     "debug-brk": "cross-env NEXT_PRIVATE_LOCAL_DEV=1 NEXT_TELEMETRY_DISABLED=1 node --inspect-brk --trace-deprecation --enable-source-maps packages/next/dist/bin/next",
     "pnpm:devPreinstall": "node scripts/create-next-bin-placeholder.mjs",
-    "postinstall": "node scripts/git-configure.mjs && node scripts/install-native.mjs",
+    "postinstall": "node scripts/git-configure.mjs && node scripts/install-native.mjs && ln -sfn .. packages/next/node_modules/next",
     "version": "pnpm install --no-frozen-lockfile && git add .",
     "prepare": "husky",
     "sync-react": "node ./scripts/sync-react.js",
@@ -137,18 +138,6 @@
     "@jest/expect-utils": "29.7.0",
     "@mdx-js/loader": "2.2.1",
     "@mdx-js/react": "2.2.1",
-    "@next/bundle-analyzer": "workspace:*",
-    "@next/env": "workspace:*",
-    "@next/eslint-plugin-next": "workspace:*",
-    "@next/font": "workspace:*",
-    "@next/mdx": "workspace:*",
-    "@next/plugin-storybook": "workspace:*",
-    "@next/polyfill-module": "workspace:*",
-    "@next/polyfill-nomodule": "workspace:*",
-    "@next/swc": "workspace:*",
-    "@next/playwright": "workspace:*",
-    "@next/third-parties": "workspace:*",
-    "@opentelemetry/api": "1.4.1",
     "@picocss/pico": "1.5.10",
     "@rspack/core": "1.6.7",
     "@slack/web-api": "7.9.1",
@@ -178,7 +167,6 @@
     "@types/trusted-types": "2.0.3",
     "@types/yargs": "16.0.11",
     "@vercel/agent-eval": "0.8.0",
-    "@vercel/devlow-bench": "workspace:*",
     "@vercel/kv": "3.0.0",
     "@vercel/og": "0.11.1",
     "abort-controller": "3.0.0",
@@ -191,7 +179,6 @@
     "cookie": "0.4.1",
     "cors": "2.8.5",
     "coveralls": "3.0.3",
-    "create-next-app": "workspace:*",
     "critters": "0.0.6",
     "cross-env": "6.0.3",
     "cross-spawn": "6.0.5",
@@ -199,7 +186,6 @@
     "es5-ext": "0.10.53",
     "escape-string-regexp": "2.0.0",
     "eslint": "9.37.0",
-    "eslint-config-next": "workspace:*",
     "eslint-formatter-codeframe": "7.32.1",
     "eslint-plugin-eslint-plugin": "5.2.1",
     "eslint-plugin-import": "2.31.0",
@@ -247,7 +233,6 @@
     "minimatch": "3.0.4",
     "moment": "^2.24.0",
     "nanoid": "3.1.30",
-    "next": "workspace:*",
     "node-fetch": "2.6.7",
     "node-plop": "0.31.1",
     "npm-run-all": "4.1.5",
