+++ packages/third-parties/package.json
@@ -26,7 +26,7 @@
     "third-party-capital": "1.0.20"
   },
   "devDependencies": {
-    "next": "16.2.0-canary.99",
+    "next": "workspace:*",
     "outdent": "0.8.0",
     "prettier": "2.5.1",
     "typescript": "5.9.2"
