+++ pnpm-workspace.yaml
@@ -6,4 +6,5 @@ packages:
   - 'turbopack/crates/*/js'
   - 'turbopack/crates/turbopack-tests/tests/execution'
   - 'turbopack/packages/*'
+  - 'test'
 updateNotifier: false
