+++ test/.gitignore
@@ -1,4 +1,4 @@
-!node_modules
+!*/**/node_modules
 .vscode
 
 e2e/**/tsconfig.json
